made more items and added funtions for attack deck to hold cards

+++ Assets/Scripts/PlayerStuff/Attacks/NonDamagingAttacks/HealWhenNoOtherAttacks.cs
@@ -15,7 +15,7 @@ public class HealWhenNoOtherAttacks : Attack
         var basicAttackData = (NonDamagingAttackData)data;
         levelData = basicAttackData.GetLevelData(level);
 
-        transform.root.GetComponentInChildren<PlayerHealth>().changeHealth(levelData.value);
+        transform.root.GetComponentInChildren<PlayerHealth>().changeHealth((int)levelData.value);
 
         level--;
     }
